requirements.txt
atk==0.1.3.8b0
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==9.4.0
pypng==0.20220715.0
q-share==0.6
qrcode==7.4.2
tk==0.1.0
typing_extensions==4.5.0
